refactor : organize station-related DTOs
