+++ src/main/java/kuit/subway/service/StationService.java
@@ -3,8 +3,8 @@ package kuit.subway.service;
 import jakarta.persistence.EntityExistsException;
 import kuit.subway.domain.Station;
 import kuit.subway.repository.StationRepository;
-import kuit.subway.response.GetStationsResponse;
-import kuit.subway.response.PostStationResponse;
+import kuit.subway.response.station.GetStationsResponse;
+import kuit.subway.response.station.PostStationResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
